chore: Bump up Zookeeper docker image tag (3.9)
